Added recipe cards to profile page and edited heroes.
Added the recipe cards for created and saved recipes to the profile. They are in a grid/list structure and tapping on one will bring you to that respective recipe's page.

Also edited the tags for heroes a little so they are based on the index of a recipe. Removes conflicting hero tags.